+++ src/transformations/d.h
@@ -80,6 +80,15 @@ namespace jags {
         bool checkParameterValue(std::vector<double const *> const &args) const;
         double evaluate(std::vector<double const *> const &args) const;
     };
+
+    class scale_d2r : public ScalarFunction 
+    {
+      public:
+        scale_d2r();
+
+        bool checkParameterValue(std::vector<double const *> const &args) const;
+        double evaluate(std::vector<double const *> const &args) const;
+    };
   }
 }